@@ -1,4 +1,4 @@
-class WeatherType < ActiveRecord::Base
+class WeatherType < ApplicationRecord
   has_many :runs
   validates :name, presence: true
 end
